add support for quoting names using backticks
This patch will add backticks as a possible quote char around
names. It will not add the possibility to us special characters
as names, as would be possible in SQL using backticks. It is just
there because a lot of schemas are created using backticks despite
there being no special chars in the name. Those schemas and queries
could not be parsed by ramsql.